apps/api/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg[binary]
alembic

python-jose
passlib[bcrypt]

pydantic
pydantic-settings
email-validator
